add constant MAX_PAGE_COUNT
One of the issues opened requested specifying max with the url.
Unfortunately because the rendered feed is cached and not the raw data,
it would be more involved to request 1000 posts from Google and cache a
feed with only a subset of the values.

Instead this always requests the same value from Google that we intend
to cache.

I found this parameter from
http://www.reddit.com/r/googleplus/comments/iq21n/getting_the_google_feed_for_any_profile_in_json/